AssertTests shared between Scala 2 and Scala 3
